+++ tests/unit/community_label_propagation.c
@@ -32,20 +32,24 @@ int main(void) {
     igraph_vector_t weights;
     igraph_vector_int_t initial;
     igraph_vector_bool_t fixed;
-    igraph_integer_t i, j;
+    igraph_integer_t i, j, k;
 
     /* Simple triangle graph, the output should be always one community */
     igraph_small(&g, 0, IGRAPH_UNDIRECTED, 0,  1,  0,  2,  1,  2, -1);
     igraph_vector_int_init(&membership, 0);
 
-    for (j = 0; j < 100; j++) {
-        /* label propagation is a stochastic method */
-        igraph_rng_seed(igraph_rng_default(), j);
+    igraph_lpa_variant_t variants[3] = {IGRAPH_LPA_DOMINANCE, IGRAPH_LPA_RETENTION, IGRAPH_LPA_FAST};
+    for (i = 0; i < 3; i++) {
+        for (j = 0; j < 100; j++) {
+            /* label propagation is a stochastic method */
+            igraph_rng_seed(igraph_rng_default(), j);
 
-        igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, 0, 0, 0);
+            igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, NULL, NULL, NULL, variants[i]);
 
-        for (i = 0; i < 3; i++)
-            IGRAPH_ASSERT(VECTOR(membership)[i] == VECTOR(membership)[0]);
+            for (k = 0; k < 3; k++) {
+                IGRAPH_ASSERT(VECTOR(membership)[k] == VECTOR(membership)[0]);
+            }
+        }
     }
 
     igraph_destroy(&g);
@@ -73,7 +77,9 @@ int main(void) {
                  31, 32, 31, 33, 32, 33,
                  -1);
 
-    igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, 0, 0, 0);
+    igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, NULL, NULL, NULL, IGRAPH_LPA_DOMINANCE);
+    igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, NULL, NULL, NULL, IGRAPH_LPA_RETENTION);
+    igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, NULL, NULL, NULL, IGRAPH_LPA_FAST);
 
     igraph_destroy(&g);
 
@@ -87,21 +93,58 @@ int main(void) {
     VECTOR(fixed)[3] = 1;
     VECTOR(fixed)[4] = 1;
     VECTOR(fixed)[5] = 1;
-    igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, &weights,
-                                       &initial, &fixed);
-    for (i = 0; i < igraph_vcount(&g); i++)
-        IGRAPH_ASSERT(VECTOR(membership)[i] == (i < 2 ? 0 : 1));
 
-    igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, 0,
-                                       &initial, &fixed);
-    for (i = 0; i < igraph_vcount(&g); i++)
-        IGRAPH_ASSERT(VECTOR(membership)[i] == 0);
+    for (i = 0; i < 3; i++) {
+        igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, &weights,
+                                           &initial, &fixed, variants[i]);
+        for (j = 0; j < igraph_vcount(&g); j++) {
+            IGRAPH_ASSERT(VECTOR(membership)[j] == (j < 2 ? 0 : 1));
+        }
 
-    /* Check whether it works with no fixed vertices at all
-     * while an initial configuration is given -- see bug
-     * #570902 in Launchpad. This is a simple smoke test only. */
+        igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, 0,
+                                           &initial, &fixed, variants[i]);
+        for (j = 0; j < igraph_vcount(&g); j++) {
+            IGRAPH_ASSERT(VECTOR(membership)[j] == 0);
+        }
+
+        /* Check whether it works with no fixed vertices at all
+        * while an initial configuration is given -- see bug
+        * #570902 in Launchpad. This is a simple smoke test only. */
+        igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, &weights,
+                                           &initial, NULL, variants[i]);
+    }
+
+    igraph_vector_bool_destroy(&fixed);
+    igraph_vector_destroy(&weights);
+    igraph_vector_int_destroy(&initial);
+    igraph_destroy(&g);
+
+    /* Test line graph with fixed and initial memberships */
+    igraph_small(&g, 3, IGRAPH_UNDIRECTED, 0, 1, 1, 2, 2, 3, -1);
+
+    igraph_vector_init(&weights, 3);
+    VECTOR(weights)[0] = 2;
+    VECTOR(weights)[1] = 1;
+    VECTOR(weights)[2] = 2;
+
+    igraph_vector_int_init(&initial, 4);
+    VECTOR(initial)[0] = 0;
+    VECTOR(initial)[1] = -1;
+    VECTOR(initial)[2] = -1;
+    VECTOR(initial)[3] = 1;
+
+    igraph_vector_bool_init(&fixed, 4);
+    VECTOR(fixed)[0] = 1;
+    VECTOR(fixed)[1] = 0;
+    VECTOR(fixed)[2] = 0;
+    VECTOR(fixed)[3] = 1;
+
     igraph_community_label_propagation(&g, &membership, IGRAPH_ALL, &weights,
-                                       &initial, 0);
+                                       &initial, &fixed, IGRAPH_LPA_DOMINANCE);
+
+    igraph_vector_int_print(&membership);
+    IGRAPH_ASSERT(VECTOR(membership)[0] == 0);
+    IGRAPH_ASSERT(VECTOR(membership)[3] == 1);
 
     igraph_vector_bool_destroy(&fixed);
     igraph_vector_destroy(&weights);